renames acceptance to integration
- like the naming a bit better
